+++ e2e/simulated_test.go
@@ -14,7 +14,6 @@ import (
 	"github.com/poteto-go/go-alchemy-sdk/constant"
 	"github.com/poteto-go/go-alchemy-sdk/deployer"
 	"github.com/poteto-go/go-alchemy-sdk/gas"
-	"github.com/poteto-go/go-alchemy-sdk/namespace"
 	"github.com/poteto-go/go-alchemy-sdk/typeddata"
 	"github.com/poteto-go/go-alchemy-sdk/types"
 	"github.com/poteto-go/go-alchemy-sdk/wallet"
@@ -1053,19 +1052,4 @@ func TestSimulated_StableCoin_FiatToken(t *testing.T) {
 
 func TestSimulated_Debug(t *testing.T) {
 	t.Skip("Debug.Snapshot / Debug.RevertTo use evm_snapshot / evm_revert over provider.Send, unavailable on simulated backend")
-}
-
-func mintERC1155(t *testing.T, contract *artifacts.ERC1155, w types.Wallet, contractHex string, transact namespace.ITransact, id, amount *big.Int) {
-	t.Helper()
-	data := contract.PackMint(common.HexToAddress(initAddress), id, amount)
-	txHash, err := w.SendTransaction(types.TransactionRequest{
-		From:     initAddress,
-		To:       contractHex,
-		Value:    "0x0",
-		GasLimit: 300000,
-		Data:     data,
-	})
-	assert.Nil(t, err)
-	_, err = transact.WaitMined(context.Background(), txHash.Hex())
-	assert.Nil(t, err)
 }
